+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'mozfet:autoform-materialize',
   summary: 'Materialize theme for Autoform',
-  version: '4.0.9',
+  version: '4.0.11',
   git: 'https://github.com/mozfet/meteor-autoform-materialize.git'
 });
 
@@ -13,7 +13,10 @@ Npm.depends({
 Package.onUse(function(api) {
   api.versionsFrom('METEOR@1.6');
   api.use(['templating@1.3.0', 'blaze@2.3.0'], 'client');
-  api.use(['aldeed:autoform@6.2.0', 'mozfet:materialize-icons@1.1.0'], 'client');
+  api.use([
+    'aldeed:autoform@6.2.0',
+    'mozfet:materialize-icons@1.1.0'
+  ], 'client');
   api.use(['ecmascript', 'underscore'], 'client');
   api.use('fourseven:scss@4.10.0');
   api.addFiles(['index.js', 'style.scss'], 'client');